Refactor metadata code
